Crate lazy_regex_proc_macros

Macros

bytes_lazy_regex

Return an instance of once_cell::sync::Lazy<bytes::Regex> that you can use in a public static declaration.

bytes_regex

Return a lazy static regex::bytes::Regex checked at compilation time and built at first use.

bytes_regex_captures

Extract captured groups as a tuple of &u8

bytes_regex_captures_iter

Returns an iterator that yields successive non-overlapping matches in the given haystack.

bytes_regex_find

Extract the leftmost match of the regex in the second argument as a &[u8]

bytes_regex_if

bytes_regex_is_match

Test whether an expression matches a lazy static bytes::Regex regular expression (the regex is checked at compile time)

bytes_regex_replace

Replaces the leftmost match in the second argument using the replacer given as third argument.

bytes_regex_replace_all

Replaces all non-overlapping matches in the second argument using the replacer given as third argument.

bytes_regex_switch

Define a set of lazy static statically compiled regexes, with a block or expression for each one. The first matching expression is computed with the named capture groups declaring &str variables available for this computation. If no regex matches, return None.

lazy_regex

Return an instance of once_cell::sync::Lazy<regex::Regex> or once_cell::sync::Lazy<regex::bytes::Regex> that you can use in a public static declaration.

regex

Return a lazy static Regex checked at compilation time and built at first use.

regex_captures

Extract captured groups as a tuple of &str.

regex_captures_iter

Returns an iterator that yields successive non-overlapping matches in the given haystack. The iterator yields values of type regex::Captures.

regex_find

Extract the leftmost match of the regex in the second argument, as a &str, or a &[u8] if the B flag is set.

regex_if

Return an Option, with T being the type returned by the block or expression given as third argument.

regex_is_match

Test whether an expression matches a lazy static regular expression (the regex is checked at compile time)

regex_replace

Replaces the leftmost match in the second argument using the replacer given as third argument.

regex_replace_all

Replaces all non-overlapping matches in the second argument using the replacer given as third argument.

regex_switch

Define a set of lazy static statically compiled regexes, with a block or expression for each one. The first matching expression is computed with the named capture groups declaring &str variables available for this computation. If no regex matches, return None.